fix(data): convert LocalDate.toEpochDays() Long to Int in weekKeyFor
toEpochDays() returns Long in this kotlinx-datetime version, not Int as
weekKeyFor's declared return type assumed — caught by CI's compile step
(:data:compileDebugKotlinAndroid).

+++ data/src/commonMain/kotlin/com/levelchef/data/repository/WeeklyChallengeRepositoryImpl.kt
@@ -128,6 +128,6 @@ class WeeklyChallengeRepositoryImpl(
         database.weeklyChallengeQueries.totalAwardedXp().executeAsOne().toInt()
 
     private companion object {
-        fun weekKeyFor(instant: Instant): Int = (instant.toLocalDateTime(TimeZone.UTC).date.toEpochDays() / 7)
+        fun weekKeyFor(instant: Instant): Int = (instant.toLocalDateTime(TimeZone.UTC).date.toEpochDays() / 7).toInt()
     }
 }
